+++ src/error/utf8_error.rs
@@ -4,6 +4,9 @@ use core::convert::Infallible;
 use core::error::Error;
 use core::fmt::{self, Display, Formatter};
 
+#[cfg(feature = "oct")]
+use oct::error::GenericDecodeError;
+
 /// An invalid UTF-8 sequence was encountered.
 #[derive(Debug, Eq, PartialEq)]
 #[must_use]
@@ -28,5 +31,19 @@ impl From<Infallible> for Utf8Error {
 	#[inline(always)]
 	fn from(_value: Infallible) -> Self {
 		unreachable!()
+	}
+}
+
+#[cfg(feature = "oct")]
+#[cfg_attr(doc, doc(cfg(feature = "oct")))]
+impl From<Utf8Error> for GenericDecodeError {
+	#[inline(always)]
+	fn from(value: Utf8Error) -> Self {
+		let e = oct::error::Utf8Error {
+			value: value.value,
+			index: value.index,
+		};
+
+		Self::BadString(e)
 	}
 }